Seperated role logic from firebase config
